Remove transferred_amount_checks
