+++ chainstate/tx-verifier/src/transaction_verifier/input_output_policy/tests/mod.rs
@@ -35,6 +35,7 @@ mod constraints_tests;
 mod homomorphism;
 mod outputs_utils;
 mod purpose_tests;
+mod reward_tests;
 
 fn get_random_inputs_combination(
     rng: &mut impl Rng,
